sources/requirements.txt
asgiref==3.5.2
asttokens==2.4.1
backcall==0.2.0
backports.zoneinfo==0.2.1
colorama==0.4.6
crispy-bootstrap5==0.7
decorator==5.1.1
Django==4.1.1
django-appconf==1.0.6
django-crispy-forms==1.14.0
django-extensions==3.2.1
django-select2==8.0.0
executing==2.1.0
ipdb==0.13.13
ipython==8.6.0
jedi==0.19.1
matplotlib-inline==0.1.7
parso==0.8.4
pickleshare==0.7.5
pillow==10.4.0
prompt_toolkit==3.0.48
pure_eval==0.2.3
pydotplus==2.0.2
Pygments==2.18.0
pyparsing==3.0.9
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.3
tomli==2.0.1
traitlets==5.14.3
tzdata==2024.2
wcwidth==0.2.13
python-dateutil
